# src/query/generated/assets/openapi/spec.yml
...
swagger: "2.0"
info:
  title: "M3DB"
  description: "A distributed time series database using M3TSZ float64 compression. You can find out more about M3DB at [http://m3db.github.io/m3/](http://m3db.github.io/m3/) or on [GitHub](https://github.com/sniperkit/snk.fork.m3)."
  version: "1.0.0"
  contact:
    name: "M3BD Google Group"
    url: "https://groups.google.com/forum/#!forum/m3db"
  license:
    name: "Apache License 2.0"
    url: "https://github.com/sniperkit/snk.fork.m3/blob/master/LICENSE"
host: "hostname:7201"
basePath: "/api/v1"
...
schemes:
- "http"
paths:
  /namespace:
    get:
      tags:
        - "namespace"
      summary: "Get namespaces"
      operationId: "namespaceGet"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/NamespaceGetResponse"
        500:
          description: ""
          schema:
            $ref: "#/definitions/GenericError"
    post:
      tags:
      - "namespace"
      summary: "Add a namespace"
      operationId: "namespaceAdd"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "body"
        in: "body"
        schema:
          $ref: "#/definitions/NamespaceAddRequest"
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/NamespaceGetResponse"
        400:
          description: ""
          schema:
            $ref: "#/definitions/GenericError"
        500:
          description: ""
          schema:
            $ref: "#/definitions/GenericError"
...
definitions:
  NamespaceAddRequest:
    type: "object"
    properties:
      name:
        type: "string"
      options:
        $ref: "#/definitions/NamespaceOptions"
  NamespaceOptions:
    type: "object"
    properties:
      bootstrapEnabled:
        type: "boolean"
      flushEnabled:
        type: "boolean"
      writesToCommitLog:
        type: "boolean"
      cleanupEnabled:
        type: "boolean"
      repairEnabled:
        type: "boolean"
      retentionOptions:
        $ref: "#/definitions/RetentionOptions"
      snapshotEnabled:
        type: "boolean"
      indexOptions:
        $ref: "#/definitions/IndexOptions"
  RetentionOptions:
    type: "object"
    properties:
      retentionPeriodNanos:
        type: "integer"
        format: "int64"
      blockSizeNanos:
        type: "integer"
        format: "int64"
      bufferFutureNanos:
        type: "integer"
        format: "int64"
      bufferPastNanos:
        type: "integer"
        format: "int64"
      blockDataExpiry:
        type: "boolean"
      blockDataExpiryAfterNotAccessPeriodNanos:
        type: "integer"
        format: "int64"
  IndexOptions:
    type: "object"
    properties:
      enabled:
        type: "boolean"
      blockSizeNanos:
        type: "integer"
        format: "int64"
  NamespaceGetResponse:
    type: "object"
    properties:
      registry:
        $ref: "#/definitions/NamespaceRegistry"
  NamespaceRegistry:
    type: "object"
    properties:
      namespaces:
        type: "object"
        additionalProperties:
          $ref: "#/definitions/NamespaceOptions"
...
  GenericError:
    type: "object"
    properties:
      error:
        type: "string"
...